+++ CHANGELOG.rst
@@ -47,6 +47,7 @@ Fixed
 
 * Fix restoring quotas bug while rally cleanup
 * Don't load compute resources in tempest verifier if nova is not enabled
+* Fix cleanup when neutron is not available
 
 [3.0.0] - 2024-05-23
 --------------------